requirements.txt
streamlit==1.40.0
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
urllib3==2.2.3
fpdf2==2.8.2
python-dateutil==2.9.0
